Rearrage files, handles duplicates, words to lower
