--- Boggle-CSCE-361/Controllers/IWordScorerController.cs
@@ -1,7 +1,0 @@
-﻿namespace Boggle_CSCE_361.Controllers
-{
-    public interface IWordScorerController
-    {
-        int getScore(String word);
-    }
-}
