+++ Boggle-CSCE-361/Controllers/Interfaces/IWordPossibilityController.cs
@@ -1,4 +1,4 @@
-﻿namespace Boggle_CSCE_361.Controllers
+﻿namespace Boggle_CSCE_361.Controllers.Interfaces
 {
     public interface IWordPossibilityController
     {
@@ -7,13 +7,13 @@
          * Uses recursion to recurse through all the possible routes from the letter.
          * Returns true if the word is possible at the starting indices. Otherwise, returns false.
          */
-        bool isWordPossibleHere(String word, String[,] grid, int startRow, int startColumn, int letterIndex);
+        bool isWordPossibleHere(string word, string[,] grid, int startRow, int startColumn, int letterIndex);
 
         /*
          * Uses the isWordPossibleHere() function.
          * Loops over the whole grid to check if the word is possible at each index.
          * If the word is found to be possible at one index, it will return true. Otherwise, returns false.
          */
-        bool isWordPossibleGrid(String[,] grid, String word);
+        bool isWordPossibleGrid(string[,] grid, string word);
     }
 }
