+++ Boggle-CSCE-361/Controllers/Interfaces/IWordScorerController.cs
@@ -1,0 +1,7 @@
+﻿namespace Boggle_CSCE_361.Controllers.Interfaces
+{
+    public interface IWordScorerController
+    {
+        int getScore(string word);
+    }
+}
