--- Boggle-CSCE-361/Controllers/WordCheck.cs
@@ -1,12 +1,0 @@
-﻿using System.Reflection.Metadata.Ecma335;
-
-namespace Boggle_CSCE_361.Controllers
-{
-    public class WordCheck
-    {
-        public static bool CheckIfWordIsCorrect(string word)
-        {
-	        return true;
-        }
-    }
-}
